requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Pillow==10.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
werkzeug==2.3.7